fixing embla carousel by making separate component for outcome step
